requirements.txt
matplotlib==3.7.1
nltk==3.8.1
numpy==1.25.0
pandas==2.0.2
pyspellchecker==0.7.2
pytest==7.3.2
PyYAML==6.0
PyYAML==6.0
rapidfuzz==3.1.1
scikit_learn==1.2.2
scipy==1.11.1
schema==0.7.5
setuptools==67.6.1
spacy==3.6.0
textblob==0.17.1
wordcloud==1.9.2
